weather_mcp_server/requirements.txt
# Weather MCP Server - Dependencias
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.24.0
python-dotenv>=1.0.0
